+++ modules/host/variables.tf
@@ -81,6 +81,7 @@ variable "ipv4_subnet_id" {
 variable "private_ipv4" {
   description = "Private IP for the server"
   type        = string
+  default     = null
 }
 
 variable "server_type" {
@@ -198,6 +199,18 @@ variable "disable_ipv6" {
   type        = bool
   default     = false
   description = "Whether to disable ipv4 on the server. If you disable ipv4 and ipv6 make sure you have an access to your private network."
+}
+
+variable "primary_ipv4_id" {
+  type        = number
+  default     = null
+  description = "Optional existing or module-managed Primary IPv4 ID to assign to the server."
+}
+
+variable "primary_ipv6_id" {
+  type        = number
+  default     = null
+  description = "Optional existing or module-managed Primary IPv6 ID to assign to the server."
 }
 
 variable "network_id" {